requirements.txt
python-telegram-bot==20.7
fastapi==0.109.0
uvicorn==0.27.0
asyncpg==0.29.0
apscheduler==3.10.4
openai==1.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.3
python-multipart==0.0.9
aiohttp==3.9.3
python-dotenv==1.0.1
